api/requirements.txt
fastapi==0.104.1
fastapi-mail==1.4.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
motor==3.3.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
jinja2==3.1.3
aiofiles==23.2.1
